Replace init greendao in Application to GreendaoHelper
