+++ app/src/main/java/cn/edu/uestc/meet_on_the_road_of_uestc/MyApplication.java
@@ -8,9 +8,11 @@ import com.tencent.tinker.loader.shareutil.ShareConstants;
 
 import cn.edu.uestc.meet_on_the_road_of_uestc.greenDao.DaoMaster;
 import cn.edu.uestc.meet_on_the_road_of_uestc.greenDao.DaoSession;
+import cn.edu.uestc.meet_on_the_road_of_uestc.greenDao.GreenDaoHelper;
 import dev.DevUtils;
 
 public class MyApplication extends TinkerApplication {
+    GreenDaoHelper greenDaoHelper=new GreenDaoHelper();
     public MyApplication(){
         super(ShareConstants.TINKER_ENABLE_ALL, "cn.edu.uestc.meet_on_the_road_of_uestc.MyApplicationLike",
                 "com.tencent.tinker.loader.TinkerLoader", false);
@@ -21,24 +23,10 @@ public class MyApplication extends TinkerApplication {
         super.onCreate();
         context = getApplicationContext();
         DevUtils.init(getApplicationContext());
-        initGreenDao();
+        greenDaoHelper.initGreenDao();
     }
     public static Context getMyContext(){
         return context;
     }
 
-    /**
-     * 初始化GreenDao,直接在Application中进行初始化操作
-     */
-    private void initGreenDao() {
-        DaoMaster.DevOpenHelper helper = new DaoMaster.DevOpenHelper(this, "moou.db");
-        SQLiteDatabase db = helper.getWritableDatabase();
-        DaoMaster daoMaster = new DaoMaster(db);
-        daoSession = daoMaster.newSession();
-    }
-
-    private DaoSession daoSession;
-    public DaoSession getDaoSession() {
-        return daoSession;
-    }
 }
