--- app/src/main/java/cn/edu/uestc/meet_on_the_road_of_uestc/help/view/HelpView.java
@@ -1,8 +1,0 @@
-package cn.edu.uestc.meet_on_the_road_of_uestc.help.view;
-
-public class HelpView implements IView{
-    @Override
-    public void showData() {
-
-    }
-}
